Add monitoring of lights, and safety semaphores

+++ C-Code/include/ManualGPIO.h
@@ -6,7 +6,7 @@
 #define IMAGETEST_MANUALGPIO_H
 int stop;
 void pinMode(int pin, short mode);
-void digitalWrite(int pin,short value);
+void digitalWrite(int pin,int value);
 int digitalRead(int pin);
 void blink(short pin, double frequency,double duration);
 int gpio_unexport(int pin);
